+++ packages/dso-toolkit/src/components/list-button/list-button.models.ts
@@ -12,6 +12,7 @@ export interface ListButton {
   min?: number;
   max?: number;
   checked?: boolean;
+  manual?: boolean;
   dsoCountChange: (e: CustomEvent<ListButtonChangeEvent>) => void;
   dsoSelectedChange: (e: CustomEvent<ListButtonSelectedEvent>) => void;
 }